data_generator/requirements.txt
faker
python-dateutil